* commit '5bb0a2cff01db0731e9d40bf3cd723b76352a16b':
  SoftAVCEncoder: fix mismatched type for comparison.

+++ media/libstagefright/codecs/avc/enc/SoftAVCEncoder.cpp
@@ -268,7 +268,7 @@ OMX_ERRORTYPE SoftAVCEncoder::initEncParams() {
 
     int32_t nMacroBlocks = divUp(mWidth, 16) * divUp(mHeight, 16);
     CHECK(mSliceGroup == NULL);
-    if (nMacroBlocks > SIZE_MAX / sizeof(uint32_t)) {
+    if ((size_t)nMacroBlocks > SIZE_MAX / sizeof(uint32_t)) {
         ALOGE("requested memory size is too big.");
         return OMX_ErrorUndefined;
     }
